+++ IO/TransferFunction2D.cpp
@@ -103,12 +103,12 @@ bool TransferFunction2D::Load(const std::string& filename) {
   m_Trans1D.Load(file, m_iSize.x);
 
   // load swatch count
-  unsigned int iSwatchCount;
+  UINT32 iSwatchCount;
   file >> iSwatchCount;
   m_Swatches.resize(iSwatchCount);
 
   // load Swatches
-  for (unsigned int i = 0;i<m_Swatches.size();i++) m_Swatches[i].Load(file);
+  for (size_t i = 0;i<m_Swatches.size();i++) m_Swatches[i].Load(file);
 
   file.close();
 
@@ -130,7 +130,7 @@ bool TransferFunction2D::Save(const std::string& filename) {
   file << m_Swatches.size() << endl;
 
   // save Swatches
-  for (unsigned int i = 0;i<m_Swatches.size();i++) m_Swatches[i].Save(file);
+  for (size_t i = 0;i<m_Swatches.size();i++) m_Swatches[i].Save(file);
 
   file.close();
 
@@ -186,7 +186,7 @@ void TransferFunction2D::GetShortArray(unsigned short** psData, unsigned short s
   RenderTransferFunction();
   unsigned short *psDataIterator = *psData;
   FLOATVECTOR4  *piSourceIterator = m_pColorData->GetDataPointer();
-  for (unsigned int i = 0;i<m_pColorData->GetSize().area();i++) {
+  for (size_t i = 0;i<m_pColorData->GetSize().area();i++) {
     *psDataIterator++ = (unsigned short)((*piSourceIterator)[0]*sUsedRange);
     *psDataIterator++ = (unsigned short)((*piSourceIterator)[1]*sUsedRange);
     *psDataIterator++ = (unsigned short)((*piSourceIterator)[2]*sUsedRange);
@@ -298,11 +298,11 @@ void TransferFunction2D::Update1DTrans(const TransferFunction1D* p1DTrans) {
   size_t iSize = min<size_t>(m_iSize.x,  m_Trans1D.GetSize());
 
   m_Trans1DImage = QImage(int(iSize), 1, QImage::Format_ARGB32);
-  for (unsigned int i = 0;i<iSize;i++) {
-    m_Trans1DImage.setPixel(i,0,qRgba(int(m_Trans1D.vColorData[i][0]*255),
-                                    int(m_Trans1D.vColorData[i][1]*255),
-                                    int(m_Trans1D.vColorData[i][2]*255),
-                                    int(m_Trans1D.vColorData[i][3]*255)));
+  for (size_t i = 0;i<iSize;i++) {
+    m_Trans1DImage.setPixel(int(i),0,qRgba(int(m_Trans1D.vColorData[i][0]*255),
+                                           int(m_Trans1D.vColorData[i][1]*255),
+                                           int(m_Trans1D.vColorData[i][2]*255),
+                                           int(m_Trans1D.vColorData[i][3]*255)));
   }
 
 }
@@ -310,12 +310,12 @@ void TransferFunction2D::Update1DTrans(const TransferFunction1D* p1DTrans) {
 // ************************************************************************************************************
 
 void TFPolygon::Load(ifstream& file) {
-  unsigned int iSize;
+  UINT32 iSize;
   file >> iSize;
   pPoints.resize(iSize);
 
-  for(unsigned int i=0;i<pPoints.size();++i){
-    for(unsigned int j=0;j<2;++j){
+  for(size_t i=0;i<pPoints.size();++i){
+    for(size_t j=0;j<2;++j){
       file >> pPoints[i][j];
     }
   }
@@ -325,9 +325,9 @@ void TFPolygon::Load(ifstream& file) {
 
   file >> iSize;
   pGradientStops.resize(iSize);
-  for(unsigned int i=0;i<pGradientStops.size();++i){
+  for(size_t i=0;i<pGradientStops.size();++i){
     file >> pGradientStops[i].first;
-    for(unsigned int j=0;j<4;++j){
+    for(size_t j=0;j<4;++j){
       file >> pGradientStops[i].second[j];
     }
   }
@@ -335,10 +335,10 @@ void TFPolygon::Load(ifstream& file) {
 }
 
 void TFPolygon::Save(ofstream& file) {
-  file << pPoints.size() << endl;
+  file << UINT32(pPoints.size()) << endl;
 
-  for(unsigned int i=0;i<pPoints.size();++i){
-    for(unsigned int j=0;j<2;++j){
+  for(size_t i=0;i<pPoints.size();++i){
+    for(size_t j=0;j<2;++j){
       file << pPoints[i][j] << " ";
     }
     file << endl;
@@ -349,9 +349,9 @@ void TFPolygon::Save(ofstream& file) {
   file << endl;
   file << pGradientStops.size() << endl;
 
-  for(unsigned int i=0;i<pGradientStops.size();++i){
+  for(size_t i=0;i<pGradientStops.size();++i){
     file << pGradientStops[i].first << "  ";
-    for(unsigned int j=0;j<4;++j){
+    for(size_t j=0;j<4;++j){
       file << pGradientStops[i].second[j] << " ";
     }
     file << endl;
